Init : Documentation 설정
- documentation 패키지 생성
- documentation name patter 변경
- doumentation 템플릿 생성
