+++ src/test/java/numble/bankingapi/documentation/BankingDocumentation.java
@@ -1,0 +1,31 @@
+package numble.bankingapi.documentation;
+
+import org.junit.jupiter.api.Test;
+
+public class BankingDocumentation extends DocumentationTemplate{
+
+	@Test
+	void getHistory() {
+
+	}
+
+	@Test
+	void deposit() {
+
+	}
+
+	@Test
+	void withdraw() {
+
+	}
+
+	@Test
+	void transfer() {
+
+	}
+
+	@Test
+	void getTargets() {
+
+	}
+}
